split PacketHandle and SessionHandle into separate files and implementations
