+++ src/System.Data.SqlClient/src/System/Data/SqlClient/TdsParserStateObject.cs
@@ -20,102 +20,6 @@ namespace System.Data.SqlClient
         internal long _value;
     }
 
-    internal readonly ref struct PacketHandle
-    {
-        public const int NativePointerType = 1;
-        public const int NativePacketType = 2;
-        public const int ManagedPacketType = 3;
-
-#if FEATURE_INTEROPSNI
-        public readonly IntPtr NativePointer;
-        public readonly SNIPacket NativePacket;
-#endif
-        public readonly SNI.SNIPacket ManagedPacket;
-        public readonly int Type;
-
-        private PacketHandle(IntPtr nativePointer,
-#if FEATURE_INTEROPSNI
-            SNIPacket
-#else
-            int
-#endif
-
-            nativePacket, SNI.SNIPacket managedPacket, int type)
-        {
-            Type = type;
-            ManagedPacket = managedPacket;
-#if FEATURE_INTEROPSNI
-            NativePointer = nativePointer;
-            NativePacket = nativePacket;
-#endif
-        }
-
-        public static PacketHandle FromManagedPacket(SNI.SNIPacket managedPacket)
-        {
-            return new PacketHandle(default, default, managedPacket, ManagedPacketType);
-        }
-
-#if FEATURE_INTEROPSNI
-        public static PacketHandle FromNativePointer(IntPtr nativePointer)
-        {
-            return new PacketHandle(nativePointer,default,default, NativePointerType);
-        }
-
-        public static PacketHandle FromNativePacket(SNIPacket nativePacket)
-        {
-            return new PacketHandle(default, nativePacket, default, NativePacketType);
-        }
-#endif
-
-
-    }
-
-    internal readonly ref struct SessionHandle
-    {
-        public const int NativeHandleType = 1;
-        public const int ManagedHandleType = 2;
-
-        public readonly SNI.SNIHandle ManagedHandle;
-#if FEATURE_INTEROPSNI
-        public readonly SNIHandle NativeHandle;
-#endif
-        public readonly int Type;
-
-        public SessionHandle(SNI.SNIHandle managedHandle,
-#if FEATURE_INTEROPSNI
-            SNIHandle
-#else
-            int
-#endif
-            nativeHandle,
-            int type
-        )
-        {
-            Type = type;
-            ManagedHandle = managedHandle;
-#if FEATURE_INTEROPSNI
-            NativeHandle = nativeHandle;
-#endif
-        }
-
-        public bool IsNull =>
-#if FEATURE_INTEROPSNI
-            (Type == NativeHandleType) ? NativeHandle == null : 
-#endif
-            ManagedHandle == null;
-
-        public static SessionHandle FromManagedSession(SNI.SNIHandle managedSessionHandle)
-        {
-            return new SessionHandle(managedSessionHandle, default, ManagedHandleType);
-        }
-#if FEATURE_INTEROPSNI
-        public static SessionHandle FromNativeHandle(SNIHandle nativeSessionHandle)
-        {
-            return new SessionHandle(default, nativeSessionHandle, NativeHandleType);
-        }
-#endif
-    }
-
     internal abstract class TdsParserStateObject
     {
         private const int AttentionTimeoutSeconds = 5;
